remove submit order functionality
now any item in a cart will be auto processed for the weekly order. This
removes the step where the user must click the "submit order" button for
their order to be processed for the week.

Almost weekly, someone would forget to press the button and we would
have to manually submit their order.