catch exception when trying to update cart inventory
If an item was placed in a cart, and then the sku of that item was
changed, an exception would be raised when updating the item inventory
when trying to update/remove that item from the cart.

we monkey patch a fix, but I think there is a possibility for the item
inventory to be slightly off now.