Add DropSite model and update ProfileForm.save() to handle it
Notification reception information is stored as JSON with the format:
    {"drop site name": {"last_hash": ""}}

last_hash is presently not used, but is there for future use.

JSON loading and dumping is handled with a property to keep code clean, but this does result in the unexpected behavior of `user.drop_site.notifications_received = {"Hollywood": {"last_hash": ""}}` causing an update rather than an assignment. I left a comment in the relevant location to clarify this.

This isn't tested, as my server's network configuration got nuked during the deploy process.