Handling for duplicate phone numbers; additional logging
When a duplicate phone number is provided, it will simply be dropped and the SIB user will be created or updated with the provided email as if it has no phone number. This is a workaround for the bizaare HTTP 200 provided by SIB in the instance of duplicate phone numbers.