fix bug in SIB dropsite list initialization
Results were limited to 10 lists, thus leading to duplicate lists being
created when the original wasn't returned in the 10 lists from the api.

I updated the limit to 50, which is the max the api allows. I also moved
the dropsite lists to another folder to keep the chances of going over
50 lists almost 0.

Also, use the dropsite folder id as a setting, thus preventing the need to fetch
all folders & match the name.